+++ src/Control/Eff/Concurrent/Protocol/Observer/Queue.hs
@@ -137,7 +137,7 @@ withObservationQueue queueLimit e = do
 spawnLinkObservationQueueWriter
   :: forall o q h
    . ( TangibleObserver o
-     , IsPdu (Observer o) 'Asynchronous
+     , HasPdu (Observer o) 'Asynchronous
      , Member Logs q
      , Lifted IO q
      , LogsTo h (Processes q)
@@ -148,7 +148,7 @@ spawnLinkObservationQueueWriter q = do
   cbo <- startLink (MkObservationQueue q)
   pure (toObserver cbo)
 
-instance (TangibleObserver o, IsPdu (Observer o) 'Asynchronous, Lifted IO q, Member Logs q) => Server (ObservationQueue o) (Processes q) where
+instance (TangibleObserver o, HasPdu (Observer o) 'Asynchronous, Lifted IO q, Member Logs q) => Server (ObservationQueue o) (Processes q) where
   type Protocol (ObservationQueue o) = Observer o
 
   data instance StartArgument (ObservationQueue o) (Processes q) =