+++ test/ProcessBehaviourTestCases.hs
@@ -91,7 +91,7 @@ data ReturnToSender
 
 type instance ToPretty ReturnToSender = PutStr "ReturnToSender"
 
-instance Typeable r => IsPdu ReturnToSender r where
+instance Typeable r => HasPdu ReturnToSender r where
  data instance Pdu ReturnToSender r where
    ReturnToSender :: ProcessId -> String -> Pdu ReturnToSender ('Synchronous Bool)
    StopReturnToSender :: Pdu ReturnToSender ('Synchronous ())